Add diffusion to the process
Add a little approximation of diffusion of photons through the glass (after they've left the phosphors). It's not really accurate but I tried to hit a balance where the effect is subtle, but visible, but also doesn't blow out the color on bright scenes. This was a good middle ground.