Break GraphicsDevice up
Working towards making this a more wrappable thing than it is, which means I want to fully separate the graphics API from the internal code. This is a step along that path - now GraphicsDevice is an interface - IGraphicsDevice - that can be filled in with whatever graphics API.
There are some oddities that I don't know how to solve - shader creation is done via enum instead of filename or file contents, in order to let the graphics API side of things handle the loading.