Reshuffle some math to fix a display bug with the shadow mask
When the texture got changed from RGBAFloat16 to RGBA8, I checked in with the comment "Not sure why this was 16, but it's not necessary. Nothing in it exceeds 1.0 and there's no extra precision requirement, so switching it to 8 bit."
Turns out that was false: values were absolutely cresting above 1.0, plus there was some additional precision loss when applying the shadowMaskStrength to the RGBA8 version of the shadow mask.

INSTEAD, the math to get the shadow mask strength applied (as well as a mysterious -0.15 bias to the mask value) is now done in RGBToCRT.hlsl instead of GenerateScreenTexture, which means *now* there are no values above 1.0 in the screen texture, nor is there any appreciable precision loss anymore.