Better screen curvature emulation
The previous screen curvature, it turns out, worked (well enough) almost entirely by accident - it was attempting a mapping from "for this uv coordinate what's the xy screen position" but it did effectively a multiply by z instead of divide by z. However, because what we actually wanted was "for this xy screen coordinate, what's the uv coordinate", that accidentally multiply-by-z ended up scaling down the visible screen which looked ... kinda correct. But it was never *quite* right.

This version, instead, basically ray casts to a centered unit sphere, scaling the spread of the cast rays based on the intended distortion (less spread for less distortion), then using an approximation of atan2 to get the uv coordinate of the hit.

(This change additionally adjusts the ranges of the distortion/mask sliders to reflect the new ranges involved as well as fixes a bug where hitting Cancel wouldn't revert the display back to where it was.