Change how settings work
Split the settings into Source Settings and other Settings - source settings being the ones that are unlikely to change (and require a rebuild of the internal components due to texture size changes). the other settings are the ones that don't tend to require any reallocation, so can be done more frequently.